working api with dockerfile

+++ diss-img-tool-lw/first_image_generator.py
@@ -30,7 +30,8 @@ def cleanup_models():
     folder = './diss-img-tool-lw/pretrained_models/high_ren'
     for filename in os.listdir(folder):
         file_path = os.path.join(folder, filename)
-        if os.path.isfile(file_path) or os.path.islink(file_path):
+        print(filename)
+        if (os.path.isfile(file_path) or os.path.islink(file_path)) and filename.endswith('.pth'):
             os.unlink(file_path)
 
 
@@ -58,7 +59,11 @@ def main(image_path_sketch, image_path_art):
     
 
 # normal main run
-cleanup_models()
+try:
+    cleanup_models()
+except Exception as e:
+        print(e)
+
 opt = bicycle_gan_opts(name)
 model = create_model(opt)
 model.setup(opt)
